+++ packages/azure-tts-client/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @ssml-builder-js/azure-tts-client
 
+## 2.7.0
+
 ## 2.6.0
 
 ### Patch Changes